+++ examples/main/modules.tf
@@ -27,6 +27,12 @@ module "logs" {
   resource_group_name = module.rg.resource_group_name
 }
 
+data "azurerm_subnet" "example" {
+  name                 = "backend"
+  virtual_network_name = "production"
+  resource_group_name  = module.rg.resource_group_name
+}
+
 module "servicebus" {
   source  = "claranet/service-bus/azurerm"
   version = "x.x.x"
@@ -39,47 +45,31 @@ module "servicebus" {
 
   resource_group_name = module.rg.resource_group_name
 
-  servicebus_namespaces_queues = {
-    # You can just create a servicebus_namespace
-    servicebus0 = {}
-
-    # Or create a servicebus_namespace with some queues with default values
-    servicebus1 = {
-      queues = {
-        queue1 = {}
-        queue2 = {}
-      }
-    }
-
-    # Or customize everything
-    servicebus2 = {
-      custom_name    = format("%s-%s-%s-custom", var.stack, var.client_name, module.azure_region.location_short)
-      sku            = "Premium"
-      capacity       = 2
-      zone_redundant = true
-
-      queues = {
-        queue100 = {
-          reader = true
-          sender = true
-          manage = true
-        }
-        queue200 = {
-          dead_lettering_on_message_expiration = true
-          default_message_ttl                  = "PT10M"
-          reader                               = true
-        }
-        queue300 = {
-          duplicate_detection_history_time_window = "PT30M"
-          sender                                  = true
-        }
-        queue400 = {
-          requires_duplicate_detection = true
-          manage                       = true
-        }
-      }
-    }
+  namespace_parameters = {
+    sku = "Premium"
   }
+
+  namespace_authorizations = {
+    listen = true
+    send   = false
+  }
+
+  # Network rules
+  network_rules_enabled    = true
+  trusted_services_allowed = true
+  allowed_cidrs = [
+    "1.2.3.4/32",
+  ]
+  subnet_ids = [
+    data.azurerm_subnet.example.id,
+  ]
+
+  servicebus_queues = [{
+    name                = "myqueue"
+    default_message_ttl = 5 # 5min
+
+    dead_lettering_on_message_expiration = true
+  }]
 
   logs_destinations_ids = [
     module.logs.logs_storage_account_id,